Slight renames for stream cache class
